refactor(example): create routing modules
