--- example-app/app/routing/hub/index.ts
@@ -1,2 +1,0 @@
-export * from './app.notes';
-export * from './app.routes';
