+++ example-app/app/views/about/about.module.ts
@@ -1,12 +1,10 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { RouterModule } from '@angular/router';
-
-import { aboutRoutes } from './hub';
 import { AboutComponent } from './components/about.component';
+import { AboutRoutingModule } from './hub/about-routing.module';
 
 @NgModule({
   declarations: [AboutComponent],
-  imports: [CommonModule, RouterModule.forChild(aboutRoutes)]
+  imports: [CommonModule, AboutRoutingModule]
 })
 export class AboutModule {}
