+++ example-app/app/views/about/hub/about-routing.module.ts
@@ -1,0 +1,15 @@
+import { NgModule } from '@angular/core';
+import { RouterModule } from '@angular/router';
+import { createFeature } from 'lib';
+import { ABOUT_NOTES_KEY, AboutNotes } from './about.notes';
+import { aboutRoutes } from './about.routes';
+
+export const aboutSlice = createFeature<AboutNotes>(aboutRoutes, {
+  key: ABOUT_NOTES_KEY
+});
+
+@NgModule({
+  imports: [RouterModule.forChild(aboutRoutes)],
+  exports: [RouterModule]
+})
+export class AboutRoutingModule {}
