--- example-app/app/views/about/hub/index.ts
@@ -1,2 +1,0 @@
-export * from './about.notes';
-export * from './about.routes';
