+++ example-app/app/views/car/hub/car.routes.ts
@@ -1,7 +1,4 @@
-import { createFeature } from 'lib';
 import { CarComponent } from '../components/carComponent';
-import { CAR_NOTES_KEY, CarNotes } from './car.notes';
-import { detailsSlice } from '../../details/hub';
 
 export const carRoutes = [
   {
@@ -15,8 +12,3 @@ export const carRoutes = [
     component: CarComponent
   }
 ];
-
-export const carSlice = createFeature<CarNotes>(carRoutes, {
-  detached: { details: detailsSlice },
-  key: CAR_NOTES_KEY
-});
