--- example-app/app/views/car/hub/index.ts
@@ -1,2 +1,0 @@
-export * from './car.notes';
-export * from './car.routes';
