--- example-app/app/views/details/hub/index.ts
@@ -1,2 +1,0 @@
-export * from './details.notes';
-export * from './details.routes';
