+++ example-app/app/views/home/hub/home.routes.ts
@@ -1,7 +1,5 @@
 import { Routes } from '@angular/router';
 import { HomeComponent } from '../components/home.component';
-import { createFeature } from 'lib';
-import { HOME_NOTES_KEY, HomeNotes } from './home.notes';
 
 export const homeRoutes: Routes = [
   {
@@ -9,7 +7,3 @@ export const homeRoutes: Routes = [
     component: HomeComponent
   }
 ];
-
-export const homeSlice = createFeature<HomeNotes>(homeRoutes, {
-  key: HOME_NOTES_KEY
-});
