--- example-app/app/views/home/hub/index.ts
@@ -1,2 +1,0 @@
-export * from './home.routes';
-export * from './home.notes';
