--- example-app/app/views/info/hub/index.ts
@@ -1,2 +1,0 @@
-export * from './info.notes';
-export * from './info.routes';
