+++ example-app/app/views/info/hub/info-routing.module.ts
@@ -1,0 +1,15 @@
+import { NgModule } from '@angular/core';
+import { RouterModule } from '@angular/router';
+import { createFeature } from 'lib';
+import { INFO_NOTES_KEY, InfoNotes } from './info.notes';
+import { infoRoutes } from './info.routes';
+
+export const infoSlice = createFeature<InfoNotes>(infoRoutes, {
+  key: INFO_NOTES_KEY
+});
+
+@NgModule({
+  imports: [RouterModule.forChild(infoRoutes)],
+  exports: [RouterModule]
+})
+export class InfoRoutingModule {}
