+++ example-app/app/views/info/hub/info.routes.ts
@@ -1,6 +1,4 @@
 import { Routes } from '@angular/router';
-import { createFeature } from 'lib';
-import { INFO_NOTES_KEY, InfoNotes } from './info.notes';
 import { InfoComponent } from '../components/info.component';
 
 export const infoRoutes: Routes = [
@@ -9,7 +7,3 @@ export const infoRoutes: Routes = [
     component: InfoComponent
   }
 ];
-
-export const infoSlice = createFeature<InfoNotes>(infoRoutes, {
-  key: INFO_NOTES_KEY
-});
