--- example-app/app/views/users/hub/index.ts
@@ -1,2 +1,0 @@
-export * from './users.routes';
-export * from './users.notes';
