+++ example-app/app/views/users/hub/users.routes.ts
@@ -1,8 +1,6 @@
 import { UsersComponent } from '../components/users.component';
 import { UserComponent } from '../components/user.component';
 import { ProfileComponent } from '../components/profile.component';
-import { createFeature } from 'lib';
-import { USERS_NOTES_KEY, UsersChildNotes, UsersNotes } from './users.notes';
 
 export const usersRoutes = [
   {
@@ -14,8 +12,3 @@ export const usersRoutes = [
     ]
   }
 ];
-
-export const usersSlice = createFeature<UsersNotes, UsersChildNotes>(
-  usersRoutes,
-  { key: USERS_NOTES_KEY }
-);
