+++ example-app/app/views/users/users.module.ts
@@ -1,14 +1,13 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { RouterModule } from '@angular/router';
 import { UsersComponent } from './components/users.component';
 import { UserComponent } from './components/user.component';
 import { ProfileComponent } from './components/profile.component';
 import { NavigationModule } from 'routeshub';
-import { usersRoutes } from './hub';
+import { UsersRoutingModule } from './hub/users-routing.module';
 
 @NgModule({
-  imports: [CommonModule, NavigationModule, RouterModule.forChild(usersRoutes)],
+  imports: [CommonModule, NavigationModule, UsersRoutingModule],
   declarations: [UsersComponent, UserComponent, ProfileComponent]
 })
 export class UsersModule {}
